new measures

+++ Project/reviews4rec/quality_content.py
@@ -135,8 +135,8 @@ for category, file_name in zip(categories, data_file_names):
             users_id.append(str(js['reviewerID']))
             items_id.append(str(js['asin']))
             ratings.append(str(js['overall']))
-            helpful_votes.append(str(int(js['helpful'][0])))
-            total_votes.append(str(int(js['helpful'][1])))
+            helpful_votes.append(str(js['helpful'][0]))
+            total_votes.append(str(js['helpful'][1]))
             # i += 1
             # if i > 10000:
             #     break
@@ -157,11 +157,11 @@ for category, file_name in zip(categories, data_file_names):
                               'flesch_kincaid', 'smog', 'pos_entropy']
     data = pd.concat([data, quality_result], axis=1)
 
+    data['total_votes'] = data['total_votes'].astype('int32')
+    data['helpful_votes'] = data['helpful_votes'].astype('int32')
     data['help_rate'] = data['helpful_votes'] / data['total_votes']
     data['help_rate'] = data['help_rate'].astype('float32')
     data['rating'] = data['rating'].astype('int8')
-    data['total_votes'] = data['total_votes'].astype('int32')
-    data['helpful_votes'] = data['helpful_votes'].astype('int32')
     data['num_chars'].fillna(0, inplace=True)
     data['num_words'].fillna(0, inplace=True)
     data['num_sentences'].fillna(0, inplace=True)
